First functional code split into separate modules with PIL
