+++ __init__.py
@@ -1,0 +1,5 @@
+#! /usr/bin/env python3
+
+import main
+if __name__ == '__main__':
+    main.run()
